requirements.txt
loguru==0.5.3
paho-mqtt==1.6.1
pynput==1.7.5
pyobjc-core==8.0
pyobjc-framework-ApplicationServices==8.0
pyobjc-framework-Cocoa==8.0
pyobjc-framework-Quartz==8.0
six==1.16.0
